Move DpadSelectionEvent out from public API
